move files from utils to dedicated resource packages
