+++ ec/ecresource/deploymentresource/enterprisesearch/v2/enterprise_search_read_test.go
@@ -184,3 +184,35 @@ func Test_readEnterpriseSearch(t *testing.T) {
 		})
 	}
 }
+
+func Test_IsEnterpriseSearchStopped(t *testing.T) {
+	type args struct {
+		res *models.EnterpriseSearchResourceInfo
+	}
+	tests := []struct {
+		name string
+		args args
+		want bool
+	}{
+		{
+			name: "started resource returns false",
+			args: args{res: &models.EnterpriseSearchResourceInfo{Info: &models.EnterpriseSearchInfo{
+				Status: ec.String("started"),
+			}}},
+			want: false,
+		},
+		{
+			name: "stopped resource returns true",
+			args: args{res: &models.EnterpriseSearchResourceInfo{Info: &models.EnterpriseSearchInfo{
+				Status: ec.String("stopped"),
+			}}},
+			want: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := IsEnterpriseSearchStopped(tt.args.res)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
