+++ ec/ecresource/deploymentresource/kibana/v2/kibana_read_test.go
@@ -173,3 +173,35 @@ func Test_ReadKibana(t *testing.T) {
 		})
 	}
 }
+
+func Test_IsKibanaResourceStopped(t *testing.T) {
+	type args struct {
+		res *models.KibanaResourceInfo
+	}
+	tests := []struct {
+		name string
+		args args
+		want bool
+	}{
+		{
+			name: "started resource returns false",
+			args: args{res: &models.KibanaResourceInfo{Info: &models.KibanaClusterInfo{
+				Status: ec.String("started"),
+			}}},
+			want: false,
+		},
+		{
+			name: "stopped resource returns true",
+			args: args{res: &models.KibanaResourceInfo{Info: &models.KibanaClusterInfo{
+				Status: ec.String("stopped"),
+			}}},
+			want: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := IsKibanaStopped(tt.args.res)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
