--- ec/ecresource/deploymentresource/utils/enrich_elasticsearch_template.go
@@ -1,59 +1,0 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package utils
-
-import (
-	"github.com/elastic/cloud-sdk-go/pkg/models"
-	"github.com/elastic/cloud-sdk-go/pkg/util/ec"
-)
-
-func EnrichElasticsearchTemplate(tpl *models.ElasticsearchPayload, templateId, version string, useNodeRoles bool) *models.ElasticsearchPayload {
-	if tpl.Plan.DeploymentTemplate == nil {
-		tpl.Plan.DeploymentTemplate = &models.DeploymentTemplateReference{}
-	}
-
-	if tpl.Plan.DeploymentTemplate.ID == nil || *tpl.Plan.DeploymentTemplate.ID == "" {
-		tpl.Plan.DeploymentTemplate.ID = ec.String(templateId)
-	}
-
-	if tpl.Plan.Elasticsearch.Version == "" {
-		tpl.Plan.Elasticsearch.Version = version
-	}
-
-	for _, topology := range tpl.Plan.ClusterTopology {
-		if useNodeRoles {
-			topology.NodeType = nil
-			continue
-		}
-		topology.NodeRoles = nil
-	}
-
-	return tpl
-}
-
-func EsResource(res *models.DeploymentTemplateInfoV2) *models.ElasticsearchPayload {
-	if res == nil || len(res.DeploymentTemplate.Resources.Elasticsearch) == 0 {
-		return &models.ElasticsearchPayload{
-			Plan: &models.ElasticsearchClusterPlan{
-				Elasticsearch: &models.ElasticsearchConfiguration{},
-			},
-			Settings: &models.ElasticsearchClusterSettings{},
-		}
-	}
-	return res.DeploymentTemplate.Resources.Elasticsearch[0]
-}
